debug

+++ saliency_wrapper.py
@@ -63,7 +63,7 @@ class SaliencyMap4F(gym.ObservationWrapper):
         :return: the observation
         """
         # call to generate saliency map here
-        frames = torch.tensor(frames).to(self.net_device)
+        frames = torch.tensor(frames).float().to(self.net_device)
         saliency_map = torch.squeeze(self.s_map(frames)).numpy()
         frame_stack = []
         for i in range(frames.shape[0]):
